+++ src/partials/header-main.html
@@ -25,7 +25,7 @@
       </ul>
     </nav>
 
-    <div class="sign-in">
+       <div class="sign-in">
       <div class="changeMode">
         <input type="checkbox" class="checkbox" id="checkbox" />
         <label for="checkbox" class="checkbox-label">
