+++ src/sass/components/_films.scss
@@ -7,10 +7,18 @@
 }
 
 .films__info {
-  margin-top: 8px;
+  margin-top: 10px;
 
   & p:first-child {
     color: $black;
+  }
+
+  @include tablet {
+    margin-top: 8px;
+  }
+
+  @include desktop {
+    margin-top: 12px;
   }
 }
 
@@ -30,17 +38,17 @@
   @media screen and (min-width: $mobile) {
     font-weight: 500;
     font-size: 12px;
-    line-height: 14px;
+    line-height: 1.17;
   }
   @media screen and (min-width: $tablet) {
     font-weight: 500;
     font-size: 12px;
-    line-height: 16px;
+    line-height: 1.33;
   }
   @media screen and (min-width: $desktop) {
     font-weight: 500;
-    font-size: 12px;
-    line-height: 16px;
+    font-size: 20px;
+    line-height: 1.2;
   }
 }
